requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
joblib
scikit-learn
pandas
xgboost
category_encoders
